[RF] Slim down `RooFitDriver.h` by moving things to other files
The RooFitDriver is central to how the new RooFit BatchMode works, and
it's interface should be small to make it accessible to new
contributors.
