+++ roofit/roofitcore/res/RooFit/BatchModeHelpers.h
@@ -23,8 +23,17 @@ class RooAbsPdf;
 class RooAbsReal;
 class RooArgSet;
 
+namespace ROOT {
+namespace Experimental {
+class RooFitDriver;
+}
+} // namespace ROOT
+
 namespace RooFit {
 namespace BatchModeHelpers {
+
+std::unique_ptr<RooAbsReal>
+makeDriverAbsRealWrapper(std::unique_ptr<ROOT::Experimental::RooFitDriver> driver, RooArgSet const &observables);
 
 std::unique_ptr<RooAbsReal> createNLL(RooAbsPdf &pdf, RooAbsData &data, std::unique_ptr<RooAbsReal> &&constraints,
                                       std::string const &rangeName, std::string const &addCoefRangeName,
@@ -43,7 +52,7 @@ private:
    TNamed const *_namePtr;
 };
 
-}
+} // namespace BatchModeHelpers
 } // namespace RooFit
 
 #endif
