+++ roofit/roofitcore/res/RooFitDriver.h
@@ -14,166 +14,49 @@
 #ifndef RooFit_RooFitDriver_h
 #define RooFit_RooFitDriver_h
 
-#include <RooAbsReal.h>
-#include <RooBatchCompute.h>
+#include <RooBatchCompute/DataKey.h>
 #include <RooGlobalFunc.h>
-
-#include <RooFit/Detail/Buffers.h>
+#include <RooHelpers.h>
 
 #include <chrono>
 #include <memory>
-#include <queue>
 #include <stack>
-#include <unordered_map>
 
 class RooAbsArg;
+class RooAbsReal;
 class RooAbsCategory;
 class RooAbsData;
-
-// Struct to temporarily change the operation mode of a RooAbsArg until it goes
-// out of scope.
-class ChangeOperModeRAII {
-public:
-   ChangeOperModeRAII(RooAbsArg *arg, RooAbsArg::OperMode opMode) : _arg{arg}, _oldOpMode(arg->operMode())
-   {
-      arg->setOperMode(opMode, /*recurse=*/false);
-   }
-   ~ChangeOperModeRAII() { _arg->setOperMode(_oldOpMode, /*recurse=*/false); }
-
-private:
-   RooAbsArg *_arg = nullptr;
-   RooAbsArg::OperMode _oldOpMode;
-};
 
 namespace ROOT {
 namespace Experimental {
 
+struct NodeInfo;
+
 class RooFitDriver {
 public:
+   ////////////////////
+   // Enums and aliases
+
    using DataSpansMap = std::map<const TNamed *, RooSpan<const double>>;
 
-   RooFitDriver(const RooAbsData &data, const RooAbsReal &topNode, RooArgSet const &normSet,
-                RooFit::BatchModeOption batchMode, std::string_view rangeName,
-                RooAbsCategory const *indexCat = nullptr);
+   //////////////////////////
+   // Public member functions
 
-   RooFitDriver(const RooAbsReal &topNode, RooArgSet const &normSet, DataSpansMap const &dataSpans,
-                RooFit::BatchModeOption batchMode);
+   RooFitDriver(const RooAbsReal &topNode, RooArgSet const &normSet,
+                RooFit::BatchModeOption batchMode = RooFit::BatchModeOption::Cpu);
+
+   void setData(RooAbsData const &data, std::string_view rangeName = "",
+                RooAbsCategory const *indexCatForSplitting = nullptr);
+   void setData(DataSpansMap const &dataSpans);
 
    ~RooFitDriver();
    std::vector<double> getValues();
    double getVal();
    RooAbsReal const &topNode() const { return _topNode; }
 
-   class RooAbsRealWrapper final : public RooAbsReal {
-   public:
-      RooAbsRealWrapper() {}
-      RooAbsRealWrapper(RooFitDriver &driver, RooArgSet const &observables, bool ownsDriver)
-         : RooAbsReal{"RooFitDriverWrapper", "RooFitDriverWrapper"}, _driver{&driver}, _ownsDriver{ownsDriver}
-      {
-         _driver->topNode().getParameters(&observables, _parameters, true);
-      }
-
-      RooAbsRealWrapper(const RooAbsRealWrapper &other, const char *name = 0)
-         : RooAbsReal{other, name}, _driver{other._driver}, _parameters{other._parameters}
-      {
-      }
-
-      ~RooAbsRealWrapper() override
-      {
-         if (_ownsDriver)
-            delete _driver;
-      }
-
-      TObject *clone(const char *newname) const override { return new RooAbsRealWrapper(*this, newname); }
-
-      double defaultErrorLevel() const override { return _driver->topNode().defaultErrorLevel(); }
-
-      bool getParameters(const RooArgSet * /*observables*/, RooArgSet &outputSet,
-                         bool /*stripDisconnected=true*/) const override
-      {
-         outputSet.add(_parameters);
-         return false;
-      }
-
-      double getValV(const RooArgSet *) const override { return evaluate(); }
-
-      void applyWeightSquared(bool flag) override
-      {
-         const_cast<RooAbsReal &>(_driver->topNode()).applyWeightSquared(flag);
-      }
-
-   protected:
-      double evaluate() const override { return _driver ? _driver->getVal() : 0.0; }
-
-   private:
-      RooFitDriver *_driver = nullptr;
-      RooArgSet _parameters;
-      bool _ownsDriver;
-   };
-
-   std::unique_ptr<RooAbsReal> makeAbsRealWrapper(RooArgSet const &observables)
-   {
-      return std::unique_ptr<RooAbsReal>{new RooAbsRealWrapper{*this, observables, false}};
-   }
-
-   // Static method to create a RooAbsRealWrapper that owns a given
-   // RooFitDriver passed by smart pointer.
-   static std::unique_ptr<RooAbsReal>
-   makeAbsRealWrapper(std::unique_ptr<RooFitDriver> driver, RooArgSet const &observables)
-   {
-      return std::unique_ptr<RooAbsReal>{new RooAbsRealWrapper{*driver.release(), observables, true}};
-   }
-
 private:
-   struct NodeInfo {
-
-      NodeInfo() {}
-
-      // No copying because of the owned CUDA pointers and buffers
-      NodeInfo(const NodeInfo &) = delete;
-      NodeInfo &operator=(const NodeInfo &) = delete;
-
-      /// Check the servers of a node that has been computed and release it's resources
-      /// if they are no longer needed.
-      void decrementRemainingClients()
-      {
-         if (--remClients == 0) {
-            buffer.reset();
-         }
-      }
-
-      RooAbsArg *absArg = nullptr;
-
-      std::unique_ptr<Detail::AbsBuffer> buffer;
-
-      cudaEvent_t *event = nullptr;
-      cudaEvent_t *eventStart = nullptr;
-      cudaStream_t *stream = nullptr;
-      std::chrono::microseconds cpuTime{0};
-      std::chrono::microseconds cudaTime{std::chrono::microseconds::max()};
-      std::chrono::microseconds timeLaunched{-1};
-      int nClients = 0;
-      int nServers = 0;
-      int remClients = 0;
-      int remServers = 0;
-      bool computeInScalarMode = false;
-      bool computeInGPU = false;
-      bool copyAfterEvaluation = false;
-      bool fromDataset = false;
-      std::size_t outputSize = 1;
-      ~NodeInfo()
-      {
-         if (event)
-            RooBatchCompute::dispatchCUDA->deleteCudaEvent(event);
-         if (eventStart)
-            RooBatchCompute::dispatchCUDA->deleteCudaEvent(eventStart);
-         if (stream)
-            RooBatchCompute::dispatchCUDA->deleteCudaStream(stream);
-      }
-   };
-
-   void init();
-   void setData(DataSpansMap const &dataSpans);
+   ///////////////////////////
+   // Private member functions
 
    double getValHeterogeneous();
    void handleIntegral(const RooAbsArg *node);
@@ -182,19 +65,12 @@ private:
    void markGPUNodes();
    void assignToGPU(const RooAbsArg *node);
    void computeCPUNode(const RooAbsArg *node, NodeInfo &info);
-
-   /// Temporarily change the operation mode of a RooAbsArg until the
-   /// RooFitDriver gets deleted.
-   void setOperMode(RooAbsArg *arg, RooAbsArg::OperMode opMode)
-   {
-      if (opMode != arg->operMode()) {
-         _changeOperModeRAIIs.emplace(arg, opMode);
-      }
-   }
-
+   void setOperMode(RooAbsArg *arg, RooAbsArg::OperMode opMode);
    void determineOutputSizes();
-
    bool isInComputationGraph(RooAbsArg const *arg) const;
+
+   ///////////////////////////
+   // Private member variables
 
    const RooFit::BatchModeOption _batchMode = RooFit::BatchModeOption::Off;
    int _getValInvocations = 0;
@@ -215,10 +91,11 @@ private:
    std::vector<double> _nonDerivedValues;
    std::stack<std::vector<double>> _vectorBuffers;
 
-   std::stack<ChangeOperModeRAII> _changeOperModeRAIIs;
-}; // end class RooFitDriver
+   // RAII structures to reset state of computation graph after driver destruction
+   std::stack<RooHelpers::ChangeOperModeRAII> _changeOperModeRAIIs;
+};
 
 } // end namespace Experimental
 } // end namespace ROOT
 
-#endif // ROO_FIT_DRIVER_H
+#endif
