+++ roofit/roofitcore/test/testRooFitDriver.cxx
@@ -18,6 +18,7 @@
 #include "RooNLLVar.h"
 #include "RooNLLVarNew.h"
 #include "RooFitDriver.h"
+#include "RooFit/BatchModeHelpers.h"
 #include "RooGaussian.h"
 #include "RooMinimizer.h"
 #include "RooFitResult.h"
@@ -83,9 +84,12 @@ TEST(testRooFitDriver, SimpleLikelihoodFit)
    resetGraph();
 
    // ...and now the new way with RooFitDriver
-   ROOT::Experimental::RooNLLVarNew nll("nll", "nll", model, *data->get(), false, "");
-   ROOT::Experimental::RooFitDriver driver(*data, nll, x, RooFit::BatchModeOption::Cpu, "");
-   auto resultBatchNew = doFit(*driver.makeAbsRealWrapper(*data->get()));
+   using namespace ROOT::Experimental;
+   RooNLLVarNew nll("nll", "nll", model, *data->get(), false, "");
+   auto driver = std::make_unique<RooFitDriver>(nll, x, RooFit::BatchModeOption::Cpu);
+   driver->setData(*data);
+   auto wrapper = RooFit::BatchModeHelpers::makeDriverAbsRealWrapper(std::move(driver), *data->get());
+   auto resultBatchNew = doFit(*wrapper);
    if (verbose) {
       std::cout << "-  batch mode fit took " << resultBatchNew.elapsedTime << " ms" << std::endl;
    }
